add ability to use an existing repo for streamlink

+++ makeportable.sh
@@ -5,10 +5,13 @@ set -e # quit on error
 STREAMLINK_PYTHON_ARCH="win32"
 STREAMLINK_PYTHON_VERSION="3.5.2"
 
-while getopts ":a:" option; do
+while getopts ":a:s:" option; do
     case $option in
         a)
             STREAMLINK_PYTHON_ARCH=$OPTARG
+            ;;
+        s)
+            STREAMLINK_REPO_DIR=$OPTARG
             ;;
         \?)
             echo "error: unknown option -$OPTARG"
@@ -49,11 +52,17 @@ mkdir -p "${dist_dir}"
 
 wget "${python_url}" -c -O "build/temp/python-${STREAMLINK_PYTHON_VERSION}-embed-${STREAMLINK_PYTHON_ARCH}.zip"
 
-# remove any old streamlink clone
-rm -rf "${streamlink_clone_dir}"
-git clone https://github.com/streamlink/streamlink.git ${streamlink_clone_dir}
+if [[ -z ${STREAMLINK_REPO_DIR} ]]; then
+    # remove any old streamlink clone
+    rm -rf "${streamlink_clone_dir}"
+    git clone https://github.com/streamlink/streamlink.git ${streamlink_clone_dir}
 
-pushd "${streamlink_clone_dir}"
+    STREAMLINK_REPO_DIR=${streamlink_clone_dir}
+fi
+
+pushd "${STREAMLINK_REPO_DIR}"
+git checkout .
+
 commit=$(git rev-parse --short HEAD)
 
 pip download --only-binary ":all:" --platform "${PYTHON_PLATFORM}" --python-version "35" --abi "cp35m" -d "${temp_dir}" "pycryptodome==3.4.3" "requests>=1.0,!=2.12.0,!=2.12.1,<3.0"
@@ -77,16 +86,16 @@ unzip -o "msvcrt_${PYTHON_PLATFORM}.zip" -d "${python_dir}"
 unzip -o "build/temp/pycryptodome*.whl" -d "${packages_dir}"
 unzip -o "build/temp/requests*.whl" -d "${packages_dir}"
 
-cp -r "${streamlink_clone_dir}/src/"* "${bundle_dir}/packages"
+cp -r "${STREAMLINK_REPO_DIR}/src/"* "${bundle_dir}/packages"
 cp "${root_dir}/streamlink-script.py" "${bundle_dir}/streamlink-script.py"
 cp "${root_dir}/streamlink.bat" "${bundle_dir}/streamlink.bat"
 cp "${root_dir}/NOTICE" "${bundle_dir}/NOTICE.txt"
 
 mkdir -p "$bundle_dir/rtmpdump" "$bundle_dir/ffmpeg"
-cp -r "${streamlink_clone_dir}/win32/rtmpdump/"* "${bundle_dir}/rtmpdump"
-cp -r "${streamlink_clone_dir}/win32/ffmpeg/"* "${bundle_dir}/ffmpeg"
-cp -r "${streamlink_clone_dir}/win32/streamlinkrc" "${bundle_dir}/streamlinkrc.default"
-cp -r "${streamlink_clone_dir}/win32/LICENSE.txt" "${bundle_dir}/LICENSE.txt"
+cp -r "${STREAMLINK_REPO_DIR}/win32/rtmpdump/"* "${bundle_dir}/rtmpdump"
+cp -r "${STREAMLINK_REPO_DIR}/win32/ffmpeg/"* "${bundle_dir}/ffmpeg"
+cp -r "${STREAMLINK_REPO_DIR}/win32/streamlinkrc" "${bundle_dir}/streamlinkrc.default"
+cp -r "${STREAMLINK_REPO_DIR}/win32/LICENSE.txt" "${bundle_dir}/LICENSE.txt"
 
 sed -i "s/^rtmpdump=.*/#rtmpdump=/g" "${bundle_dir}/streamlinkrc.default"
 sed -i "s/^ffmpeg-ffmpeg=.*/#ffmpeg-ffmpeg=/g" "${bundle_dir}/streamlinkrc.default"
